+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/Dynatrace/openkit-java/compare/v2.2.0...HEAD)
 
+### Added
+- User-Agent header for http requests
 ### Changed
 - Default maximum age of an entry in the beacon cache changed from 1h 45m to 45m
 - Removed AppMon functionality